Change db model, calculate score on update, etc...
This makes the db model less-performant as it has to load every single vote per post. This approach is also quite convoluted as I need to remember to calculate the score and upvoted/downvoted status properly. I'll probably change this back and clean up the code somehow but for now it's okay.